redesigns the server layout
The previous layout was a consequence of the quick symlink hack
we did when Rails 3 was out. Until then Rails only published
the documentation of the last stable release, and since people
needed still to check the documentation of Rails 2 as a quick
fix I symlinked v2.3.something in the document root.

With the recent revamp we start to fully support generation of
all releases in all branches and this layout better reflects
that. In addition, it decouples the particular API and guides
document roots (which change from time to time), from the
nginx configuration.

See the code comment in the docs generator for further details.