big refactor
The motivation to start this refactor was to rewrite configuration via
inheritance.

When I wrote that code it seemed like a bet that could make sense, but
the evolution of the project has not validated that bet:

* In practice there are more configuration points and exceptions than expected
  which has resulted in an unexpect number of classes and hops in the linear
  runtime you want to understand.

* Some targets did not inherit from the previous one, but had to jump further
  back to leverage the shared bits.

* The code that instantiated target generators was doing some dynamic dances
  which are too meta seen in perspective.

In summary, you look at this stuff as it is today, and it looks too complicated
for what it does. In particular, if a new release or master needs new code, what
you are inheriting and what to touch where is not as clear as it could be.

So, I have rewritten all this with lovely `if`s. Now, you open the generators
and see what is going on easily.

The key for this alternative is being able to compare version numbers easily and
in a readable way. There is a new `VersionNumber` class that supports comparisons
with strings (in addition to other version numbers), so you can write

    if version_number < '4'
      ...
    end

or

    if version_number <= '3.2.0'
      ...
    end

and it works as the code suggests.

Since I was on it, I have also renamed the concept of "stable" docs and such to
"release". Stable is a name that comes from the old times in which Rails only had
the stable docs online, that is, only the docs of the current stable release were
published at a time (manually). Now we publish the documentation of all modern
"releases", that is a better name.