patches documents.yaml in some v5.1.x versions
Doing integration tests for every single patch like this is proving to
be too expensive. We are going to switch to a lighter approach by
checking the modifications we expect in the repo, and assuming those
modifications are the ones that allow the docs to be built.

References https://github.com/rails/rails/issues/29865.