Disable mobi generation for rails guides on edge
The mobi generation takes a long time and it executes once for every
documentation generation that during a day can be a lot of times.

Because of this the documentation server is always running the mobi
generation taking 100% of CPU time in a 1 CPU machine. This means that
the server doesn't have many resources to do what it should be doing,
serving the documentation.

Since the edge guide change often enough that a mobi downloaded today
may have wrong information tomorrow it is better to only generate the
mobi in a stable release.