Preserve version on Autoloading Guide redirect
The v4.2 and v5.0 guides are the only guides which link to
`constant_autoloading_and_reloading.html`.  However, visiting that path
in either version currently redirects to the stable version of
`autoloading_and_reloading_constants.html`, which is a Zeitwerk guide as
of Rails 6.0.

This commit preserves the guide version when redirecting.