Insert edge badge on index.html as well
The 'edge_badge.png' is inserted in all html pages in classes and files.
Now that index.html renders a html body instead of frames, the edge
badge should be inserted as well.

Since we can insert it in all .html files, we can use `Dir.glob` to loop
over the files and remove the extra loop and extension conditional.