Clone only the tag needed to generate the docs
Before, for each released version, we were cloning the entire Rails
repository in a new directory, then checking out the tag we wanted.
This is inefficient in both time and space.

Our server is full of copies of the Rails repository, one per released
version, and each copy contains the full history of the repository,
using gigabytes of disk space.

To save in both time and space, we can clone only the specific tag we
need, using git's `--branch` and `--depth 1` options.